Solution to Check Leap Year

+++ Check Leap Year@@6788dd63490637cfc44e1b90/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    double a;
-    scanf("%lf",&a);
+    int a;
+    scanf("%d",&a);
     if(a%4==0 && a%400==0){
         printf(Leap Year);
     }else{
